+++ pom.xml
@@ -1,9 +1,9 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
 	<modelVersion>4.0.0</modelVersion>
-	<groupId>teracy</groupId>
-	<artifactId>teracy-selenium</artifactId>
-	<version>1.0</version>
+	<groupId>org.teracy.scats</groupId>
+	<artifactId>scats-core</artifactId>
+	<version>0.1.0-SNAPSHOT</version>
 	<name>teracy selenium test</name>
 	<packaging>jar</packaging>
 	<description>TERACY SELENIUM TEST</description>
