+++ src/main/java/org/teracy/scats/core/actions/MainAction.java
@@ -1,4 +1,4 @@
-package teracy.coreactions;
+package org.teracy.scats.core.actions;
 
 import java.io.File;
 
@@ -7,10 +7,10 @@ import org.openqa.selenium.Keys;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.support.ui.Select;
+import org.teracy.scats.utils.TestLogger;
 
 import net.thucydides.core.annotations.Step;
 import net.thucydides.core.steps.ScenarioSteps;
-import teracy.utils.TestLogger;
 
 public class MainAction extends ScenarioSteps {
 	/**
