+++ src/main/java/org/teracy/scats/utils/FileUtil.java
@@ -1,4 +1,4 @@
-package teracy.utils;
+package org.teracy.scats.utils;
 
 import java.io.File;
 import java.io.FileWriter;
