+++ src/main/java/org/teracy/scats/utils/RandomData.java
@@ -1,4 +1,4 @@
-package teracy.utils;
+package org.teracy.scats.utils;
 
 import java.util.Random;
